Move log level test to setup ui tests (#33)
* move log level test to setup ui

* rename setup tests
