+++ example/lib/testapps/sample_for_setup_tests.dart
@@ -1,14 +1,15 @@
+import 'package:didomi_sdk_example/widgets/initialize_small.dart';
 import 'package:didomi_sdk_example/widgets/set_log_level.dart';
+import 'package:didomi_sdk_example/widgets/setup_ui.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-void main() => runApp(SampleForLogLevelTestsApp());
+void main() => runApp(SampleForSetupUITestsApp());
 
-class SampleForLogLevelTestsApp extends StatelessWidget {
-  // This widget is the root of your application.
+class SampleForSetupUITestsApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) => MaterialApp(
-        title: "Log Level Tests",
+        title: "Dialogs Tests",
         home: HomePage(),
       );
 }
@@ -19,6 +20,8 @@ class HomePage extends StatelessWidget {
         child: ListView(
           key: Key("components_list"),
           children: [
+            InitializeSmall(),
+            SetupUI(),
             SetLogLevel(),
           ],
         ),
